restructured api handler files
